# services/learning-service/src/main/java/com/itshaharcha/learning/controller/DocController.java
package com.itshaharcha.learning.controller;

import com.itshaharcha.common.web.ApiResponse;
import com.itshaharcha.learning.dto.request.DocReadInput;
import com.itshaharcha.learning.dto.response.DocResponse;
import com.itshaharcha.learning.dto.response.PageResponse;
import com.itshaharcha.learning.service.DocService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Pageable;
import org.springframework.data.web.PageableDefault;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import java.util.UUID;

@Tag(name = "docs", description = "Documentation / reading material")
@RestController
@RequestMapping("/api/v1/learning/docs")
@RequiredArgsConstructor
public class DocController {

    private final DocService docService;

    @Operation(summary = "List documentation / reading material")
    @GetMapping
    public ApiResponse<PageResponse<DocResponse>> listDocs(
            @RequestParam(required = false) String topic,
            @PageableDefault(size = 20) Pageable pageable) {
        return ApiResponse.ok(docService.listDocs(topic, pageable));
    }

    @Operation(summary = "Record that a doc was read")
    @PostMapping("/{docId}:read")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void recordRead(@PathVariable UUID docId,
                           @Valid @RequestBody(required = false) DocReadInput input) {
        docService.recordRead(docId, input);
    }
}