# specs/learning.openapi.yaml
...
openapi: 3.1.0
info:
  title: IT-Shaharcha — Learning Service
  version: 1.0.0
  description: IT skill courses, tutorials, reading docs, and typing practice.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
    description: Versioned platform base path (behind the gateway).
security:
  - bearerAuth: []
...
paths:
...
  /learning/tracks:
    get:
      operationId: listTracks
      summary: List skill tracks.
      tags: [catalog]
      security: []
      parameters:
        - $ref: "./shared/common.yaml#/components/parameters/PageParam"
        - $ref: "./shared/common.yaml#/components/parameters/SizeParam"
        - name: q
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: A page of tracks.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "./shared/common.yaml#/components/schemas/Page"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/Track" }
        "400": { $ref: "./shared/common.yaml#/components/responses/BadRequest" }
  /learning/courses:
    get:
      operationId: listCourses
      summary: List courses, optionally filtered by track.
      tags: [catalog]
      security: []
      parameters:
        - name: trackId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: level
          in: query
          required: false
          schema: { $ref: "#/components/schemas/Level" }
        - $ref: "./shared/common.yaml#/components/parameters/PageParam"
        - $ref: "./shared/common.yaml#/components/parameters/SizeParam"
        - $ref: "./shared/common.yaml#/components/parameters/SortParam"
      responses:
        "200":
          description: A page of courses.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "./shared/common.yaml#/components/schemas/Page"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/Course" }
        "400": { $ref: "./shared/common.yaml#/components/responses/BadRequest" }
  /learning/courses/{courseId}:
    parameters:
      - name: courseId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getCourse
      summary: Get a course with its modules and lessons.
      tags: [catalog]
      security: []
      responses:
        "200":
          description: The course detail.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/CourseDetail" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }

  # ==========================================================================
  # Enrollment & lesson progress
  # ==========================================================================
...
  /learning/docs:
    get:
      operationId: listDocs
      summary: List programming documentation / reading material.
      tags: [docs]
      security: []
      parameters:
        - name: topic
          in: query
          required: false
          schema: { type: string }
        - $ref: "./shared/common.yaml#/components/parameters/PageParam"
        - $ref: "./shared/common.yaml#/components/parameters/SizeParam"
      responses:
        "200":
          description: A page of docs.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "./shared/common.yaml#/components/schemas/Page"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/Doc" }
        "400": { $ref: "./shared/common.yaml#/components/responses/BadRequest" }
  /learning/docs/{docId}:read:
    parameters:
      - name: docId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: recordDocRead
      summary: Record that a doc was read. Emits learning.doc.read.
      tags: [docs]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                durationSeconds: { type: integer, minimum: 0 }
                scrollPercent: { type: number, minimum: 0, maximum: 100 }
      responses:
        "204": { description: Read recorded. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }

  # ==========================================================================
  # Typing practice
  # ==========================================================================
...
  /learning/admin/sources:
    get:
      operationId: listContentSources
      summary: List registered content sources.
      tags: [admin]
      parameters:
        - name: type
          in: query
          required: false
          schema: { $ref: "#/components/schemas/SourceType" }
        - $ref: "./shared/common.yaml#/components/parameters/PageParam"
        - $ref: "./shared/common.yaml#/components/parameters/SizeParam"
      responses:
        "200":
          description: A page of content sources.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "./shared/common.yaml#/components/schemas/Page"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/ContentSource" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
    post:
      operationId: createContentSource
      summary: Register a content source.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentSourceInput" }
      responses:
        "201": { $ref: "#/components/responses/ContentSourceOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
  /learning/admin/sources/{sourceId}:
    parameters:
      - name: sourceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getContentSource
      summary: Get a content source.
      tags: [admin]
      responses:
        "200": { $ref: "#/components/responses/ContentSourceOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
    patch:
      operationId: updateContentSource
      summary: Update a content source (e.g. enable/disable, change schedule).
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentSourceInput" }
      responses:
        "200": { $ref: "#/components/responses/ContentSourceOk" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "422": { $ref: "./shared/common.yaml#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteContentSource
      summary: Delete a content source.
      tags: [admin]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
  /learning/admin/sources/{sourceId}:sync:
    parameters:
      - name: sourceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: syncContentSource
      summary: Trigger an immediate sync of a content source.
      tags: [admin]
      parameters:
        - $ref: "./shared/common.yaml#/components/parameters/IdempotencyKeyHeader"
      responses:
        "202":
          description: Sync started.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SourceSyncRun" }
        "401": { $ref: "./shared/common.yaml#/components/responses/Unauthorized" }
        "403": { $ref: "./shared/common.yaml#/components/responses/Forbidden" }
        "404": { $ref: "./shared/common.yaml#/components/responses/NotFound" }
        "409": { $ref: "./shared/common.yaml#/components/responses/Conflict" }
components:
  securitySchemes:
    bearerAuth:
      $ref: "./shared/common.yaml#/components/securitySchemes/bearerAuth"
  schemas:
    Level:
      type: string
      enum: [beginner, intermediate, advanced]
    Track:
      type: object
      required: [id, title]
      properties:
        id: { type: string, format: uuid, example: "33333333-3333-3333-3333-333333333333" }
        title: { type: string, example: "Web Development" }
        slug: { type: string, example: "web-development" }
        description: { type: [string, "null"], example: "From HTML basics to full-stack apps." }
        courseCount: { type: integer, minimum: 0, example: 12 }
    Course:
      type: object
      required: [id, title, level]
      properties:
        id: { type: string, format: uuid, example: "44444444-4444-4444-4444-444444444444" }
        trackId: { type: [string, "null"], format: uuid, example: "33333333-3333-3333-3333-333333333333" }
        title: { type: string, example: "JavaScript Fundamentals" }
        slug: { type: string, example: "javascript-fundamentals" }
        summary: { type: [string, "null"], example: "Variables, functions, and the DOM." }
        level: { $ref: "#/components/schemas/Level" }
        lessonCount: { type: integer, minimum: 0, example: 24 }
        estimatedMinutes: { type: [integer, "null"], minimum: 0, example: 480 }
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            modules:
              type: array
              items: { $ref: "#/components/schemas/Module" }
    Module:
      type: object
      required: [id, title, lessons]
      properties:
        id: { type: string, format: uuid, example: "55555555-5555-5555-5555-555555555555" }
        title: { type: string, example: "Functions & Scope" }
        order: { type: integer, minimum: 0, example: 2 }
        lessons:
          type: array
          items: { $ref: "#/components/schemas/Lesson" }
    Lesson:
      type: object
      required: [id, title]
      properties:
        id: { type: string, format: uuid, example: "66666666-6666-6666-6666-666666666666" }
        title: { type: string, example: "Arrow Functions" }
        order: { type: integer, minimum: 0, example: 3 }
        kind:
          type: string
          enum: [reading, video, exercise, quiz]
          default: reading
          example: video
        estimatedMinutes: { type: [integer, "null"], minimum: 0, example: 15 }
...
    Doc:
      type: object
      required: [id, title]
      properties:
        id: { type: string, format: uuid, example: "99999999-9999-9999-9999-999999999999" }
        title: { type: string, example: "HTTP Status Codes Reference" }
        topic: { type: [string, "null"], example: "web" }
        url: { type: [string, "null"], format: uri, example: "https://docs.itsh.dev/http-status-codes" }
        estimatedMinutes: { type: [integer, "null"], minimum: 0, example: 8 }
...
    SourceType:
      type: string
      enum: [rss, youtube, sitemap, api, manual]
    SourceTarget:
      type: string
      description: Which catalog the source populates.
      enum: [tutorials, docs]
    SourceStatus:
      type: string
      enum: [active, paused, error]
    ContentSourceInput:
      type: object
      required: [name, type, target, url]
      properties:
        name: { type: string, example: "freeCodeCamp YouTube" }
        type: { $ref: "#/components/schemas/SourceType" }
        target: { $ref: "#/components/schemas/SourceTarget" }
        url: { type: string, format: uri, description: Feed/endpoint URL to poll., example: "https://www.youtube.com/feeds/videos.xml?channel_id=UC8butISFwT-Wl7EV0hUK0BQ" }
        enabled: { type: boolean, default: true, example: true }
        schedule:
          type: [string, "null"]
          description: Cron expression for periodic sync; null = manual only.
          example: "0 6 * * *"
        defaultTopic:
          type: [string, "null"]
          description: Topic tag applied to items ingested from this source.
          example: "frontend"
    ContentSource:
      allOf:
        - $ref: "#/components/schemas/ContentSourceInput"
        - type: object
          required: [id, status, createdAt]
          properties:
            id: { type: string, format: uuid, example: "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb" }
            status: { $ref: "#/components/schemas/SourceStatus" }
            createdAt: { type: string, format: date-time, example: "2026-05-01T09:00:00Z" }
            lastSyncedAt: { type: [string, "null"], format: date-time, example: "2026-05-29T06:00:00Z" }
            lastError: { type: [string, "null"], example: null }
            itemCount:
              type: integer
              minimum: 0
              default: 0
              description: Items currently ingested from this source.
              example: 134
    SourceSyncRun:
      type: object
      required: [sourceId, runId, status]
      properties:
        sourceId: { type: string, format: uuid, example: "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb" }
        runId: { type: string, format: uuid, example: "cccccccc-cccc-cccc-cccc-cccccccccccc" }
        status:
          type: string
          enum: [queued, running, completed, failed]
          example: queued
        startedAt: { type: [string, "null"], format: date-time, example: "2026-05-29T10:15:00Z" }
        itemsImported: { type: [integer, "null"], minimum: 0, example: null }

  # --------------------------------------------------------------------------
  # Reusable typed success responses for the admin authoring endpoints.
  # --------------------------------------------------------------------------
  responses:
...
    ContentSourceOk:
      description: The content source.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "./shared/common.yaml#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data: { $ref: "#/components/schemas/ContentSource" }